Refactor DAP state machine + fix SWJ_Pins
Before the refactor both `Jtag` and `Swd` were assuming the ownership
of the `Context`. Now, all three are expected to know how to convert
from one another which should better support the use case where
internal fields are less dynamic type-wise (type-states, SPI used
in SWD mode that has to be deconstructed on a switch and own the
pins etc.).

Now, DAP state is supposed to be changed lazily; that is only when
currently executed command requires a different mode/state, the switch
will occur. This enables the correct implementation of the SWJ_Pins
command that is supposed to maintain the requested state of the pins
until another command arrives.